@@ -10,10 +10,10 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/elastic/elastic-agent/internal/pkg/agent/application/coordinator"
 	"github.com/elastic/elastic-agent/internal/pkg/fleetapi"
 	"github.com/elastic/elastic-agent/internal/pkg/fleetapi/acker"
 	"github.com/elastic/elastic-agent/pkg/core/logger"
+	"github.com/elastic/elastic-agent/pkg/features"
 )
 
 // Upgrade is a handler for UPGRADE action.
@@ -21,17 +21,20 @@ import (
 // from repository specified by fleet.
 type Upgrade struct {
 	log        *logger.Logger
-	coord      *coordinator.Coordinator
+	coord      upgradeCoordinator
 	bkgActions []fleetapi.Action
 	bkgCancel  context.CancelFunc
 	bkgMutex   sync.Mutex
+
+	tamperProtectionFn func() bool // allows to inject the flag for tests, defaults to features.TamperProtection
 }
 
 // NewUpgrade creates a new Upgrade handler.
-func NewUpgrade(log *logger.Logger, coord *coordinator.Coordinator) *Upgrade {
+func NewUpgrade(log *logger.Logger, coord upgradeCoordinator) *Upgrade {
 	return &Upgrade{
-		log:   log,
-		coord: coord,
+		log:                log,
+		coord:              coord,
+		tamperProtectionFn: features.TamperProtection,
 	}
 }
 
@@ -51,6 +54,25 @@ func (h *Upgrade) Handle(ctx context.Context, a fleetapi.Action, ack acker.Acker
 	if !runAsync {
 		return nil
 	}
+
+	if h.tamperProtectionFn() {
+		// Find inputs that want to receive UPGRADE action
+		// Endpoint needs to receive a signed UPGRADE action in order to be able to uncontain itself
+		state := h.coord.State()
+		ucs := findMatchingUnitsByActionType(state, a.Type())
+		if len(ucs) > 0 {
+			h.log.Debugf("handlerUpgrade: proxy/dispatch action '%+v'", a)
+			err := notifyUnitsOfProxiedAction(ctx, h.log, action, ucs, h.coord.PerformAction)
+			h.log.Debugf("handlerUpgrade: after action dispatched '%+v', err: %v", a, err)
+			if err != nil {
+				return err
+			}
+		} else {
+			// Log and continue
+			h.log.Debugf("No components running for %v action type", a.Type())
+		}
+	}
+
 	go func() {
 		h.log.Infof("starting upgrade to version %s in background", action.Version)
 		if err := h.coord.Upgrade(asyncCtx, action.Version, action.SourceURI, action, false); err != nil {
