@@ -188,7 +188,7 @@ func installCmd(streams *cli.IOStreams, cmd *cobra.Command) error {
 
 		defer func() {
 			if err != nil {
-				_ = install.Uninstall(cfgFile, topPath)
+				_ = install.Uninstall(cfgFile, topPath, "")
 			}
 		}()
 
@@ -222,7 +222,7 @@ func installCmd(streams *cli.IOStreams, cmd *cobra.Command) error {
 		if err != nil {
 			if status != install.PackageInstall {
 				var exitErr *exec.ExitError
-				_ = install.Uninstall(cfgFile, topPath)
+				_ = install.Uninstall(cfgFile, topPath, "")
 				if err != nil && errors.As(err, &exitErr) {
 					return fmt.Errorf("enroll command failed with exit code: %d", exitErr.ExitCode())
 				}
