@@ -141,7 +141,7 @@ func TestManager_SimpleComponentErr(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -233,7 +233,7 @@ func TestManager_FakeInput_StartStop(t *testing.T) {
 							subErrCh <- fmt.Errorf("unit failed: %s", unit.Message)
 						} else if unit.State == client.UnitStateHealthy {
 							// remove the component which will stop it
-							err := m.Update([]component.Component{})
+							err := m.Update(component.Model{Components: []component.Component{}})
 							if err != nil {
 								subErrCh <- err
 							}
@@ -256,7 +256,7 @@ func TestManager_FakeInput_StartStop(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -380,7 +380,7 @@ func TestManager_FakeInput_Features(t *testing.T) {
 							Fqdn: &proto.FQDNFeature{Enabled: true},
 						}
 
-						err := m.Update([]component.Component{comp})
+						err := m.Update(component.Model{Components: []component.Component{comp}})
 						if err != nil {
 							subscriptionErrCh <- fmt.Errorf("[case %d]: failed to update component: %w",
 								healthIteration, err)
@@ -435,7 +435,7 @@ func TestManager_FakeInput_Features(t *testing.T) {
 						"message": "Fake Healthy",
 					})
 
-					err := m.Update([]component.Component{comp})
+					err := m.Update(component.Model{Components: []component.Component{comp}})
 					if err != nil {
 						t.Logf("error updating component state to health: %v", err)
 
@@ -455,7 +455,7 @@ func TestManager_FakeInput_Features(t *testing.T) {
 	defer drainErrChan(managerErrCh)
 	defer drainErrChan(subscriptionErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	timeout := 30 * time.Second
@@ -571,7 +571,7 @@ func TestManager_FakeInput_BadUnitToGood(t *testing.T) {
 							}
 
 							unitBad = false
-							err := m.Update([]component.Component{updatedComp})
+							err := m.Update(component.Model{Components: []component.Component{updatedComp}})
 							if err != nil {
 								subErrCh <- err
 							}
@@ -599,7 +599,7 @@ func TestManager_FakeInput_BadUnitToGood(t *testing.T) {
 								}
 							} else if unit.State == client.UnitStateHealthy {
 								// bad unit is now healthy; stop the component
-								err := m.Update([]component.Component{})
+								err := m.Update(component.Model{Components: []component.Component{}})
 								if err != nil {
 									subErrCh <- err
 								}
@@ -623,7 +623,7 @@ func TestManager_FakeInput_BadUnitToGood(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -738,7 +738,7 @@ func TestManager_FakeInput_GoodUnitToBad(t *testing.T) {
 									Err:  errors.New("hard-error for config"),
 								}
 								unitGood = false
-								err := m.Update([]component.Component{updatedComp})
+								err := m.Update(component.Model{Components: []component.Component{updatedComp}})
 								if err != nil {
 									subErrCh <- err
 								}
@@ -751,7 +751,7 @@ func TestManager_FakeInput_GoodUnitToBad(t *testing.T) {
 						} else {
 							if unit.State == client.UnitStateFailed {
 								// went to failed; stop whole component
-								err := m.Update([]component.Component{})
+								err := m.Update(component.Model{Components: []component.Component{}})
 								if err != nil {
 									subErrCh <- err
 								}
@@ -773,7 +773,7 @@ func TestManager_FakeInput_GoodUnitToBad(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -877,7 +877,7 @@ func TestManager_FakeInput_NoDeadlock(t *testing.T) {
 			}
 			i += 1
 			comp = updatedComp
-			err := m.Update([]component.Component{updatedComp})
+			err := m.Update(component.Model{Components: []component.Component{updatedComp}})
 			if err != nil {
 				updatedErr <- err
 				return
@@ -918,7 +918,7 @@ LOOP:
 		case <-endTimer.C:
 			// no deadlock after timeout (all good stop the component)
 			updatedCancel()
-			_ = m.Update([]component.Component{})
+			_ = m.Update(component.Model{Components: []component.Component{}})
 			break LOOP
 		case err := <-errCh:
 			require.NoError(t, err)
@@ -1010,7 +1010,7 @@ func TestManager_FakeInput_Configure(t *testing.T) {
 								"state":   int(client.UnitStateDegraded),
 								"message": "Fake Degraded",
 							})
-							err := m.Update([]component.Component{comp})
+							err := m.Update(component.Model{Components: []component.Component{comp}})
 							if err != nil {
 								subErrCh <- err
 							}
@@ -1033,7 +1033,7 @@ func TestManager_FakeInput_Configure(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -1151,7 +1151,7 @@ func TestManager_FakeInput_RemoveUnit(t *testing.T) {
 						} else if unit1.State == client.UnitStateHealthy {
 							// unit1 is healthy lets remove it from the component
 							comp.Units = comp.Units[0:1]
-							err := m.Update([]component.Component{comp})
+							err := m.Update(component.Model{Components: []component.Component{comp}})
 							if err != nil {
 								subErrCh <- err
 							}
@@ -1186,7 +1186,7 @@ func TestManager_FakeInput_RemoveUnit(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -1310,7 +1310,7 @@ func TestManager_FakeInput_ActionState(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -1445,7 +1445,7 @@ func TestManager_FakeInput_Restarts(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -1562,7 +1562,7 @@ func TestManager_FakeInput_Restarts_ConfigKill(t *testing.T) {
 									"message": "Fake Healthy",
 									"kill":    rp[1],
 								})
-								err := m.Update([]component.Component{comp})
+								err := m.Update(component.Model{Components: []component.Component{comp}})
 								if err != nil {
 									subErrCh <- err
 								}
@@ -1587,7 +1587,7 @@ func TestManager_FakeInput_Restarts_ConfigKill(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(1 * time.Minute)
@@ -1701,7 +1701,7 @@ func TestManager_FakeInput_KeepsRestarting(t *testing.T) {
 									"message":          fmt.Sprintf("Fake Healthy %d", lastStoppedCount),
 									"kill_on_interval": true,
 								})
-								err := m.Update([]component.Component{comp})
+								err := m.Update(component.Model{Components: []component.Component{comp}})
 								if err != nil {
 									subErrCh <- err
 								}
@@ -1729,7 +1729,7 @@ func TestManager_FakeInput_KeepsRestarting(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(1 * time.Minute)
@@ -1844,7 +1844,7 @@ func TestManager_FakeInput_RestartsOnMissedCheckins(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -1961,7 +1961,7 @@ func TestManager_FakeInput_InvalidAction(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -2158,7 +2158,7 @@ func TestManager_FakeInput_MultiComponent(t *testing.T) {
 	defer drainErrChan(subErrCh1)
 	defer drainErrChan(subErrCh2)
 
-	err = m.Update(components)
+	err = m.Update(component.Model{Components: components})
 	require.NoError(t, err)
 
 	count := 0
@@ -2314,7 +2314,7 @@ func TestManager_FakeInput_LogLevel(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update([]component.Component{comp})
+	err = m.Update(component.Model{Components: []component.Component{comp}})
 	require.NoError(t, err)
 
 	endTimer := time.NewTimer(30 * time.Second)
@@ -2528,7 +2528,7 @@ func TestManager_FakeShipper(t *testing.T) {
 									subErrCh <- err
 								} else {
 									// successful; turn it all off
-									err := m.Update([]component.Component{})
+									err := m.Update(component.Model{Components: []component.Component{}})
 									if err != nil {
 										subErrCh <- err
 									}
@@ -2557,7 +2557,7 @@ func TestManager_FakeShipper(t *testing.T) {
 									subErrCh <- err
 								} else {
 									// successful; turn it all off
-									err := m.Update([]component.Component{})
+									err := m.Update(component.Model{Components: []component.Component{}})
 									if err != nil {
 										subErrCh <- err
 									}
@@ -2592,7 +2592,7 @@ func TestManager_FakeShipper(t *testing.T) {
 									subErrCh <- err
 								} else {
 									// successful; turn it all off
-									err := m.Update([]component.Component{})
+									err := m.Update(component.Model{Components: []component.Component{}})
 									if err != nil {
 										subErrCh <- err
 									}
@@ -2617,7 +2617,7 @@ func TestManager_FakeShipper(t *testing.T) {
 	defer drainErrChan(errCh)
 	defer drainErrChan(subErrCh)
 
-	err = m.Update(comps)
+	err = m.Update(component.Model{Components: comps})
 	require.NoError(t, err)
 
 	timeout := 2 * time.Minute
@@ -2822,7 +2822,7 @@ func TestManager_FakeInput_OutputChange(t *testing.T) {
 	}
 
 	time.Sleep(100 * time.Millisecond)
-	err = m.Update(components)
+	err = m.Update(component.Model{Components: components})
 	require.NoError(t, err)
 
 	updateSleep := 300 * time.Millisecond
@@ -2831,7 +2831,7 @@ func TestManager_FakeInput_OutputChange(t *testing.T) {
 		updateSleep = time.Second
 	}
 	time.Sleep(updateSleep)
-	err = m.Update(components2)
+	err = m.Update(component.Model{Components: components2})
 	require.NoError(t, err)
 
 	count := 0
