@@ -276,7 +276,84 @@ func TestBatch(t *testing.T) {
 		},
 	}
 
-	actual, err := DetermineBatches("testdata", "batch_test")
+	actual, err := DetermineBatches("testdata", "", "batch_test")
 	require.NoError(t, err)
 	require.EqualValues(t, expected, actual)
+}
+
+var testLinuxLocalTests = []string{"TestLinuxLocal"}
+
+var testLinuxLocalBatch = []Batch{
+	{
+		OS: OS{
+			Type: "linux",
+			Arch: "amd64",
+		},
+		Tests: []BatchPackageTests{
+			{
+				Name:  "github.com/elastic/elastic-agent/pkg/testing/define/testdata",
+				Tests: testLinuxLocalTests,
+			},
+		},
+	},
+	{
+		OS: OS{
+			Type: "linux",
+			Arch: "arm64",
+		},
+		Tests: []BatchPackageTests{
+			{
+				Name:  "github.com/elastic/elastic-agent/pkg/testing/define/testdata",
+				Tests: testLinuxLocalTests,
+			},
+		},
+	},
+}
+
+func TestGoTestFlags(t *testing.T) {
+	testcases := []struct {
+		name     string
+		flags    string
+		expected []Batch
+	}{
+		{
+			name:     "Run single test",
+			flags:    "-run ^TestLinuxLocal$",
+			expected: testLinuxLocalBatch,
+		},
+		{
+			name:     "Run single test with short flag",
+			flags:    "-run ^TestLinuxLocal$ -short",
+			expected: testLinuxLocalBatch,
+		},
+		{
+			name:     "specify non-existing test",
+			flags:    "-run ^thisdoesnotexist$",
+			expected: nil,
+		},
+		{
+			name:     "specify multiple run flags - last one wins - no test",
+			flags:    "-run ^TestLinuxLocal$ -run ^thisdoesnotexist$",
+			expected: nil,
+		},
+		{
+			name:     "specify multiple run flags - last one wins - TestLinuxLocal",
+			flags:    "-run ^thisdoesnotexist$ -run ^TestLinuxLocal$",
+			expected: testLinuxLocalBatch,
+		},
+		{
+			name:     "count flag will not multiply the test entries in each batch",
+			flags:    "-run ^TestLinuxLocal$ -count 2",
+			expected: testLinuxLocalBatch,
+		},
+	}
+
+	for _, tc := range testcases {
+		t.Run(tc.name, func(t *testing.T) {
+			actual, err := DetermineBatches("testdata", tc.flags, "batch_test")
+			require.NoError(t, err)
+			require.EqualValues(t, tc.expected, actual)
+		})
+	}
+
 }
