@@ -52,7 +52,7 @@ type OSRunnerResult struct {
 // OSRunner provides an interface to run the tests on the OS.
 type OSRunner interface {
 	// Prepare prepares the runner to actual run on the host.
-	Prepare(ctx context.Context, c *ssh.Client, logger Logger, arch string, goVersion string, repoArchive string, buildPath string) error
+	Prepare(ctx context.Context, c *ssh.Client, logger Logger, arch string, goVersion string, repoArchive string, buildPaths []string) error
 	// Run runs the actual tests and provides the result.
 	Run(ctx context.Context, verbose bool, c *ssh.Client, logger Logger, agentVersion string, prefix string, batch define.Batch, env map[string]string) (OSRunnerResult, error)
 }
@@ -264,7 +264,7 @@ func (r *Runner) runMachine(ctx context.Context, sshAuth ssh.AuthMethod, logger 
 	}
 
 	// ensure that we have all the requirements for the stack if required
-	var env map[string]string
+	env := map[string]string{}
 	if batch.Batch.Stack != nil {
 		ch, err := r.getCloudForBatchID(batch.ID)
 		if err != nil {
@@ -276,17 +276,18 @@ func (r *Runner) runMachine(ctx context.Context, sshAuth ssh.AuthMethod, logger 
 			return OSRunnerResult{}, fmt.Errorf("cannot continue because stack never became ready")
 		}
 		logger.Logf("Will continue, stack is ready")
-		env = map[string]string{
-			"ELASTICSEARCH_HOST":     resp.ElasticsearchEndpoint,
-			"ELASTICSEARCH_USERNAME": resp.Username,
-			"ELASTICSEARCH_PASSWORD": resp.Password,
-			"KIBANA_HOST":            resp.KibanaEndpoint,
-			"KIBANA_USERNAME":        resp.Username,
-			"KIBANA_PASSWORD":        resp.Password,
-		}
-		logger.Logf("Created Stack with Kibana host %s, %s/%s", resp.KibanaEndpoint, resp.Username, resp.Password)
+		env["ELASTICSEARCH_HOST"] = resp.ElasticsearchEndpoint
+		env["ELASTICSEARCH_USERNAME"] = resp.Username
+		env["ELASTICSEARCH_PASSWORD"] = resp.Password
+		env["KIBANA_HOST"] = resp.KibanaEndpoint
+		env["KIBANA_USERNAME"] = resp.Username
+		env["KIBANA_PASSWORD"] = resp.Password
 
+		logger.Logf("Created Stack with Kibana host %s, %s/%s", resp.KibanaEndpoint, resp.Username, resp.Password)
 	}
+
+	// set the go test flags
+	env["GOTEST_FLAGS"] = r.cfg.TestFlags
 
 	// run the actual tests on the host
 	prefix := fmt.Sprintf("%s-%s-%s-%s", batch.LayoutOS.OS.Type, batch.LayoutOS.OS.Arch, batch.LayoutOS.OS.Distro, strings.Replace(batch.LayoutOS.OS.Version, ".", "", -1))
@@ -303,9 +304,11 @@ func (r *Runner) validate() error {
 	var requiredFiles []string
 	for _, b := range r.batches {
 		if !b.Skip {
-			buildPath := r.getBuildPath(b)
-			if !slices.Contains(requiredFiles, buildPath) {
-				requiredFiles = append(requiredFiles, buildPath)
+			buildPaths := r.getBuildPath(b)
+			for _, buildPath := range buildPaths {
+				if !slices.Contains(requiredFiles, buildPath) {
+					requiredFiles = append(requiredFiles, buildPath)
+				}
 			}
 		}
 	}
@@ -325,7 +328,7 @@ func (r *Runner) validate() error {
 }
 
 // getBuildPath returns the path of the build required for the test.
-func (r *Runner) getBuildPath(b LayoutBatch) string {
+func (r *Runner) getBuildPath(b LayoutBatch) []string {
 	arch := b.LayoutOS.OS.Arch
 	if arch == define.AMD64 {
 		arch = "x86_64"
@@ -334,7 +337,10 @@ func (r *Runner) getBuildPath(b LayoutBatch) string {
 	if b.LayoutOS.OS.Type == define.Windows {
 		ext = "zip"
 	}
-	return filepath.Join(r.cfg.BuildDir, fmt.Sprintf("elastic-agent-%s-%s-%s.%s", r.cfg.AgentVersion, b.LayoutOS.OS.Type, arch, ext))
+	hashExt := ".sha512"
+	packageName := filepath.Join(r.cfg.BuildDir, fmt.Sprintf("elastic-agent-%s-%s-%s.%s", r.cfg.AgentVersion, b.LayoutOS.OS.Type, arch, ext))
+	packageHashName := packageName + hashExt
+	return []string{packageName, packageHashName}
 }
 
 // prepare prepares for the runner to run.
